Rhythm/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.31.0
alembic>=1.13.1
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
bcrypt>=5.0.0
python-jose[cryptography]>=3.3.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
httpx>=0.27.0
